feat: adds css loader
