+++ packages/webpack-loader/css-loader/README.md
@@ -1,0 +1,1 @@
+# @compiled/webpack-loader/css-loader
